adding intValue

+++ src/main/java/com/citigroup/liquifi/autopilot/helper/AutoPilotConfiguration.java
@@ -84,7 +84,7 @@ public class AutoPilotConfiguration {
 				symbol = symbol.replaceAll("\\[.*\\]", replacement);
 			}
 			
-			defaultSymbolMap.put(entry.getKey(), symbol);
+			defaultSymbolMap.put(entry.getKey().intValue(), symbol);
 		}
 	}
